weather-microservices/api-gateway/requirements.txt
Flask==3.0.3
requests==2.32.3
prometheus-flask-exporter==0.23.1
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-instrumentation-flask==0.48b0
opentelemetry-instrumentation-requests==0.48b0
pytest==8.3.2
